Fix typings and lint
